requirements.txt
Babel==2.1.1
Django==1.8.7
Pillow==3.0.0
argparse==1.2.1
django-ajaxerrors==1.1
django-allauth==0.24.1
django-appconf==1.0.1
django-autoslug==1.9.3
django-cacheops==2.4.3
django-compressor==1.6
django-crispy-forms==1.5.2
django-datatable-view==0.8.3
django-debug-toolbar==1.4
django-mptt==0.7.4
django-phonenumber-field==1.0.0
django-polymorphic==0.7.2
django-suit==0.2.15
django-suit-redactor==0.0.2
funcy==1.6
oauthlib==1.0.3
phonenumbers==7.1.1
psycopg2==2.6.1
pyinotify==0.9.6
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
redis==2.10.5
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
sqlparse==0.1.18
wsgiref==0.1.2
